Swap DEIMS variable 'label' for variable 'name' content for EML rendering
the DEIMS variable field "Name" is the most prominent (visible) attribute field in Drupal, and makes sense that is a human readable string.  The DEIMS variable field "Label" is the 'label' used in the header of the data source column/row, which may or may not be human readable.  However, EML in LTER is interpreted as <attributeName> being populated by the 'Label', and the <attributeLabel> being populated by the human readable DEIMS 'Name'.